fix(levm): fix problems related to 32 bit architecture (#4104)
**Motivation**

- Find errors on 32 bit architecture

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- `u256_to_usize` now halts if the number doesn't fit in 32 bits instead
of 64.
- Memory size cost is computed using u64 instead of usize. The previous
implementation overflowed on 32 bit architectures in some edge cases.
Also I made it prettier :D


How this was tested: Adapting ethrex to run on 32 bit x86 in branch
[replay_32_bit_x86_testing](https://github.com/lambdaclass/ethrex/tree/replay_32_bit_x86_testing)
and running EFTests, both state and blockchain. After these changes they
all pass.
<!-- Link to issues: Resolves #111, Resolves #222 -->

Makes progress towards #4081

+++ crates/vm/levm/src/constants.rs
@@ -15,7 +15,7 @@ pub const EMPTY_CODE_HASH: H256 = H256([
     0xe5, 0x00, 0xb6, 0x53, 0xca, 0x82, 0x27, 0x3b, 0x7b, 0xfa, 0xd8, 0x04, 0x5d, 0x85, 0xa4, 0x70,
 ]);
 
-pub const MEMORY_EXPANSION_QUOTIENT: usize = 512;
+pub const MEMORY_EXPANSION_QUOTIENT: u64 = 512;
 
 // Dedicated gas limit for system calls according to EIPs 2935, 4788, 7002 and 7251
 pub const SYS_CALL_GAS_LIMIT: u64 = 30000000;
